This update adds several new properties to the MONTHCAL control and one new read-only property to the GRID control.  It includes changes to the documentation in the DOC folder.
The new properties in MONTHCAL are WIDTH, HEIGHT, and properties for all 6 colors in the control.

The new property in GRID is ROWSPERPAGE, the number of grid rows currently visible in the control.

There are two new demos and one updated demo to demonstrate these changes.

The updated Controls\MonthCalendar\MONTHCAL4 demo now demonstrates getting and setting of the new properties.

A new demo, Controls\Grid\Grid_Rows_Per_Page, demonstrates the new ROWSPERPAGE property.

Another new demo, Controls\DatePicker\DATEPICKER_2, shows how to mimic a calendar datepicker with an UPDOWN DATEPICKER, a BUTTON, and a MONTHCALENDAR.